standardize key name for postcal codes
Postal codes are currently using a number of variations in key names.
This makes it hard for library users to find information reliably in the
returned data strucutre.

zipcode is a terminology that's used in a small handful of countries,
whereas postal code is a general-use terminology that can be applied
anywhere, so let's standadize on this for the key name.

NOTE: this introduces an ABI breaking change.

Closes: #17