update whois sample and expected structure for google.cl
Tests are currently failing on google.cl since the expected dates don't
contain the timezone even though it is present in the sample data.

We'll also take a more up to date output from whois so that we can check
that the organization name is present in the data structure.