configure github workflows to run unit tests on each commit
In order to avoid situations like before the three previous commits
where tests are broken without it being visible, we should configure CI
to automatically run tests upon all incoming commits.

This also adds the use of flake8 to lint the code in order to keep a
sound code style.