fix .eu parser
The regexes used for the .eu parser don't match most of the times. They
need to use a more generic way of skipping lines until the expected
label prefixing the value. It's too bad that most of the values are not
being tested by unit tests though.

The parser for the name servers is not doing what it should: a long
string containing \n and spaces is returned. Also the "Keys:" section
for DNSSEC is included in the string. The end of section delimiter is
two blank lines, but since this can happen later in the file, we need to
non-greedily match. Then the lines that get matched need to be split and
stripped. Again, the value of the parsed name_servers is not getting
tested correctly by the unit tests

Making the tests verify values for more fields will need to be done in a
future commit since it'll require us to fix many other tests for other
TLDs.