.ve split nameservers into a list of items
This is a better user interface.

The current sample for the .ve domain does not correspond at all to what
I'm getting if I run a whois request on the domain. So there is a good
chance that we'll need to update the sample and completely rework the
regular expressions. This is more work than what I'm hoping to achieve
here (namely fixing tests)