Remove imports of builtins
There doesn't seem to be a good reason to import builtins since we're
not redefining any of the builtin functions/types.

This has a good chance to be breaking python 2.x support. But since 2.x
has been EOL'ed for a while now we should probably not ensure support
with those versions of the language.