Standardize test assertions for sample files
The test_com_allsamples test is the only one that's using self.fail()
instead of functions from the unittest.UnitTest class. Because of this,
failures show a super obscure output that does not show the actual
nature of the failure, which makes it really difficult to debug.

With self.assert*() at least when there is a difference in values, it
will get shown, leading to easier debugging.